// src/types/user.ts
export interface User {
  id: number;
  name: string;
  email: string;
  avatar: string;
  status: string;
  totalOrders: number;
  totalSpent: number;
  role: string;
  phoneNumber: string;
  country: string;
  state: string;
  city: string;
  address: string;
  zipCode: string;
  company: string;
}